Fix all review findings from 6-agent parallel review
Critical fixes:
- ipset6_added_entry adjacency check now guards against IPV6_ADDR_MAX
  overflow (previously "ffff:...:ffff" + "::" merged to "::/0")
- str2netaddr6 prefix parsing uses strtol instead of atoi (previously
  "/abc" silently became /0, expanding input to entire IPv6 space)
- IPv4 mode now converts ::ffff:x.x.x.x mapped addresses back to IPv4
- IPv6 lines in IPv4 mode are dropped gracefully with one-warning
  summary instead of hard failure (ipv6_dropped_in_ipv4_mode)
- IPv4 loader detects v2 binary header and produces clear error
- CMakeLists.txt updated with all 16 new IPv6 source files

High fixes:
- configure.ac checks for __uint128_t availability (fails with
  clear message on platforms that lack it)

Medium fixes:
- unique_ips saturates at IPV6_ADDR_MAX instead of wrapping to 0
  for full address space ranges
- --min-prefix and --prefixes now work in IPv6 mode (1..128 range)
- --default-prefix skipped in IPv6 mode (always uses /128)
- --has-ipv6 added to --help output
- MODE_ defines moved to iprange.h (shared, no duplication)
- Overflow guards added to ipset6_exclude and ipset6_diff for
  hi+1 at IPV6_ADDR_MAX boundary
- unsigned long -> size_t in set operation index variables
- UTF-8 BOM stripped from first line in IPv6 loader
- unique_ips < entries check added to binary v2 loader

Low fixes:
- Dead ipset6_chain_append removed from iprange.c
- NULL check added to ipset6_free
- Debug logging added to ipset6_added_entry

Tests: 93 total (77 IPv4 + 16 IPv6), all passing.
5 new regression tests for the critical bugs found.

+++ src/ipset6_common.c
@@ -4,7 +4,7 @@
 
 inline ipset6 *ipset6_common(ipset6 *ips1, ipset6 *ips2) {
     ipset6 *ips;
-    unsigned long int n1, n2, i1 = 0, i2 = 0;
+    size_t n1, n2, i1 = 0, i2 = 0;
     ipv6_addr_t lo1, lo2, hi1, hi2, lo, hi;
 
     if(unlikely(!(ips1->flags & IPSET_FLAG_OPTIMIZED)))
